+++ src/simulator/event.rs
@@ -1,8 +1,8 @@
-use crate::core::types::NetId;
+use crate::core::types::EntityId;
 use crate::core::types::Logic;
 
 pub struct Event {
     pub time: usize,
-    pub net: NetId,
+    pub entity: EntityId,
     pub new_value: Logic,
 }